pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Train an agent to produce a timeline of events"
authors = ["Hugo Sousa <hmosousa@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^3.0.1"
networkx = "^3.4.1"
transformers = "^4.45.2"
torch = "2.4.1"
python-dotenv = "^1.0.1"
accelerate = "^1.0.1"
fire = "^0.7.0"
omegaconf = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.0"
pre-commit = "^4.0.1"
snakeviz = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
